refactor: move CommitInformation test to its own file
