+++ tests/unit/core/scan/test_commit.py
@@ -6,7 +6,6 @@ import pytest
 
 from ggshield.core.filter import init_exclusion_regexes
 from ggshield.core.scan import Commit
-from ggshield.core.scan.commit_information import CommitInformation
 from ggshield.utils.git_shell import Filemode
 from tests.conftest import is_windows
 from tests.repository import Repository
@@ -410,59 +409,3 @@ def test_from_staged(tmp_path):
 
     paths_and_modes = [(x.path, x.filemode) for x in files]
     assert paths_and_modes == [(Path("NEW.md"), Filemode.NEW)]
-
-
-@pytest.mark.parametrize(
-    ("patch", "expected"),
-    [
-        (
-            """Author: ezra <ezra@lothal.sw>
-Date: Thu Sep 29 15:55:41 2022 +0000
-
-    Make changes
-
-ghost.txt\0cat.py\0""",
-            CommitInformation(
-                "ezra",
-                "ezra@lothal.sw",
-                "Thu Sep 29 15:55:41 2022 +0000",
-                [Path(x) for x in ("ghost.txt", "cat.py")],
-            ),
-        ),
-        # This can happen, see: https://github.com/sqlite/sqlite/commit/981706534.patch
-        (
-            """Author: emptymail <>
-Date: Thu Sep 29 15:55:41 2022 +0000
-
-    Delete gone.txt, rename old.txt to new.txt
-
-gone.txt\0new.txt\0""",
-            CommitInformation(
-                "emptymail",
-                "",
-                "Thu Sep 29 15:55:41 2022 +0000",
-                [Path(x) for x in ("gone.txt", "new.txt")],
-            ),
-        ),
-        (
-            """Author: ezra <ezra@lothal.sw>
-Date: Thu Sep 29 15:55:41 2022 +0000
-
-    An empty commit
-""",
-            CommitInformation(
-                "ezra",
-                "ezra@lothal.sw",
-                "Thu Sep 29 15:55:41 2022 +0000",
-                [],
-            ),
-        ),
-    ],
-)
-def test_commit_information_from_patch_header(patch: str, expected: CommitInformation):
-    """
-    GIVEN a patch header
-    WHEN parsing it with CommitInformation.from_patch_header()
-    THEN it extracts the expected values
-    """
-    assert CommitInformation.from_patch_header(patch) == expected
